Add files via upload
This ATM transaction project is created using files in c language .Here mainscrre is main program and by using the extern keyword in C extends the visibility of variables and functions across multiple source files like deposit, pinwrong, withdraw, etc... In the case of functions, the extern keyword is used implicitly. But with variables, have to use the keyword explicitly in the files. It is like a replica of atm transaction which is quiet similar to atm machine.